Docs and style cleanup (#30)
* Rewrapped all comment lines to 80 characters/row, and trimmed trailing whitespace.

This commit contains formatting changes only.  Only the following
changes were made:
- All comments that didn't contain rust code (basically, plain English
  explanations) were wrapped such that the right most character in a
  line was at column number <= 80.
- All trailing whitespace was removed from each line.

* Called 'cargo +nightly fmt --all' at root of crate.

This is yet another formatting change.  Kept separate to make it
easier to pull and look at the changes.

* feat: Added a code of conduct.

This is a copy of https://www.contributor-covenant.org/version/2/0/code_of_conduct.html.
No modifications have been done.

* feat: Updated the CONTRIBUTING.md file.

This is just a first pass at this, it will likely require more tweaks.

* feat: Added language on what feature requests need to have.

* fix: derive/tests/simple.rs failed to compile because of 'Field' ambiguity.

derive/tests/simple.rs uses a pair of glob imports, both of which define
'Field'.  This ambiguity was caught by the compiler, and caused a compilation
failure when running all tests in the workspace from the root.  This commit
fixes the bug.

* style: Added .rustfmt.toml to root, and used it to format everything.

This is my personal style file I use for formatting rust code.  You may
wish to change it to match your own style.

* style: Manually rewrapped various documentation comments.

Rewrapped the documentation comments to 80 columns where it made
sense to do so.

* style: Manually rewrapped doc comments.

* docs: Minor cleanup and rewording of some of the docs.