WIP: Started the 'intrusive_collections_examples' crate.
This crate is very much a work in progress.  My goal is to build a
series of examples showing off how the 'generic-field-projection'
crate can be used to compose highly complex data structures from much
simpler ones, and to do so in an efficient and safe manner.

Because some of this is *highly* experimental, it **should NOT** be
merged into master until significant testing and thinking about it
have been done.